backend - all except messages/notifications done

+++ server/src/post/post.service.ts
@@ -1,6 +1,11 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common'
+import {
+	Injectable,
+	InternalServerErrorException,
+	NotFoundException
+} from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { In, Repository } from 'typeorm'
+import { MediaService } from '../media/media.service'
 import { RelationshipService } from '../relationship/relationship.service'
 import { CreatePostDto } from './dto/create-post.dto'
 import { UpdatePostDto } from './dto/update-post.dto'
@@ -11,7 +16,8 @@ export class PostService {
 	constructor(
 		@InjectRepository(PostEntity)
 		private readonly postRepository: Repository<PostEntity>,
-		private readonly relationshipService: RelationshipService
+		private readonly relationshipService: RelationshipService,
+		private readonly mediaService: MediaService
 	) {}
 
 	async getCurrentUserPosts(userId: number): Promise<PostEntity[]> {
@@ -21,8 +27,15 @@ export class PostService {
 			},
 			relations: {
 				user: true,
-				likes: true,
-				comments: true
+				likes: {
+					user: true
+				},
+				comments: {
+					user: true,
+					replies: {
+						user: true
+					}
+				}
 			},
 			order: {
 				createdAt: 'DESC'
@@ -35,26 +48,43 @@ export class PostService {
 
 		await this.relationshipService
 			.getFriends(userId)
-			.then(friends =>
-				friends.map(friend => relatedIds.push(friend.fromUser.id))
+			.then(relations =>
+				relations.map(relation => relatedIds.push(relation.fromUser.id))
 			)
 
 		await this.relationshipService
 			.getSubscribes(userId)
-			.then(subscribes =>
-				subscribes.map(subscribe => relatedIds.push(subscribe.toUser.id))
+			.then(relations =>
+				relations.map(relation => relatedIds.push(relation.toUser.id))
 			)
 
 		return await this.postRepository.find({
 			where: {
-				user: {
-					id: In(relatedIds)
+				user: In(relatedIds)
+			},
+			relations: {
+				user: true,
+				likes: {
+					user: true
+				},
+				comments: {
+					user: true,
+					replies: true
 				}
 			}
 		})
 	}
 
 	async removePost(postId: number): Promise<boolean> {
+		const isExist = await this.postRepository.findOne({
+			where: {
+				id: postId
+			}
+		})
+
+		if (!isExist)
+			throw new NotFoundException('Попытка удаления несуществующего поста')
+
 		const isRemoved = await this.postRepository.delete({ id: postId })
 
 		if (!isRemoved)
@@ -64,13 +94,17 @@ export class PostService {
 	}
 
 	async updatePost(postId: number, dto: UpdatePostDto): Promise<boolean> {
-		const postToUpdate = await this.postRepository.find({
-			where: { id: postId }
+		const isExist = await this.postRepository.findOne({
+			where: {
+				id: postId
+			}
 		})
+		if (!isExist)
+			throw new NotFoundException('Попытка обновления несуществующего поста')
 
 		const isUpdated = await this.postRepository.update(
-			{ ...dto },
-			{ id: postId }
+			{ id: postId },
+			{ ...dto }
 		)
 
 		if (!isUpdated)
